backend/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
python-decouple==3.8
requests==2.31.0
django-cors-headers==4.4.0
